+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -253,6 +253,7 @@ sub build_isf_public_interface_contract {
             't/1197-isf-latency-clause-boundary.t',
             't/1198-isf-update-clause-boundary.t',
             't/1199-isf-shift-clause-boundary.t',
+            't/1200-isf-assemble-clause-boundary.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',